feat(login): smooth vector clock, AA text, fix gradient math
Major rendering quality improvements:
- Replace pixelated bitmap-scaled clock with smooth 7-segment vector
  digits drawn with rounded rectangles + neon glow halos
- Ghost segments show inactive LCD glass segments for realism
- Add anti-aliased fringe rendering for scaled text (username, date)
  that draws semi-transparent edge pixels at exposed block boundaries
- Fix draw_gradient_v unsigned integer underflow bug: use signed int
  arithmetic so color channels can decrease without wrapping to garbage
  (this was causing the bright green/yellow AUTHENTICATE button)
- Add WiFi arc icon, signal bars, and secure shield status icons
- Smooth bullet dots with glow halos in password field
- Lock icon in password field instead of crude pixel art shield
- Button now renders correct blue gradient (0x28,0x78,0xC0 -> 0x1C,0x5C,0x98)

+++ gui/draw.c
@@ -241,17 +241,18 @@ void draw_gradient_v(canvas_t* c, int x, int y, int w, int h,
                      uint32_t color_top, uint32_t color_bottom)
 {
     if (h <= 0) return;
-    uint32_t r0 = (color_top  >> 16) & 0xFF;
-    uint32_t g0 = (color_top  >>  8) & 0xFF;
-    uint32_t b0 =  color_top         & 0xFF;
-    uint32_t r1 = (color_bottom >> 16) & 0xFF;
-    uint32_t g1 = (color_bottom >>  8) & 0xFF;
-    uint32_t b1 =  color_bottom        & 0xFF;
+    /* Use signed arithmetic so colors can decrease without wrapping */
+    int r0 = (int)((color_top  >> 16) & 0xFF);
+    int g0 = (int)((color_top  >>  8) & 0xFF);
+    int b0 = (int)( color_top         & 0xFF);
+    int r1 = (int)((color_bottom >> 16) & 0xFF);
+    int g1 = (int)((color_bottom >>  8) & 0xFF);
+    int b1 = (int)( color_bottom        & 0xFF);
 
     for (int row = 0; row < h; row++) {
-        uint32_t r = r0 + (r1 - r0) * row / h;
-        uint32_t g = g0 + (g1 - g0) * row / h;
-        uint32_t b = b0 + (b1 - b0) * row / h;
+        int r = r0 + (r1 - r0) * row / h;
+        int g = g0 + (g1 - g0) * row / h;
+        int b = b0 + (b1 - b0) * row / h;
         draw_rect(c, x, y + row, w, 1, rgb((uint8_t)r,(uint8_t)g,(uint8_t)b));
     }
 }
